+++ frontend/src/components/markdown/render.ts
@@ -1,0 +1,19 @@
+import { unified } from "unified"
+import remarkParse from "remark-parse"
+import remarkGfm from "remark-gfm"
+import remarkRehype from "remark-rehype"
+import rehypeHighlight from "rehype-highlight"
+import rehypeStringify from "rehype-stringify"
+
+const RenderMarkdown = async (content: string) => {
+  const html = await unified()
+    .use(remarkParse)
+    .use(remarkGfm)
+    .use(remarkRehype)
+    .use(rehypeHighlight, { ignoreMissing: true })
+    .use(rehypeStringify)
+    .process(content)
+  return html.toString()
+}
+
+export default RenderMarkdown
